Wire up the download worker properly when re-hydrating
Would be nice if we could combine hydration with creating the object... I'll leave that as a future exercise. Could do it be saying all events are required to be serializable or something, then any non-serializable objects get added in a hydration step when creating the panel. May have other implications though, and this change is already quite large.